refactor(threads): useThreadItemEvents 拆 delta 队列与 normalized 管线子 hook
- 2522→1726 行：useRealtimeDeltaQueue 396 + useNormalizedRealtimePipeline 584
- flushPendingRealtimeEvents 聚合入口与 unmount flush 逐字留主文件；分段前 drain 语义未触
- 25 段字节级校验 + 11 deps 数组逐字；对外 3 导出面不变
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；useThreadItemEvents 8 文件 79 passed；useThreadEventHandlers 68 passed；gate 绿
（子代理 ItemEventsPhase2 执行，主代理复核验证）
